# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: ReqFlow Procurement Workflow API
servers:
  - url: /api
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, password]
              properties:
                name:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "401":
          description: Invalid name or password
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      responses:
        "200":
          description: Currently logged-in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "401":
          description: Not logged in

  # ── Companies ───────────────────────────────────────────────────────────────
...
  /requisitions/{id}/items:
    get:
      operationId: listRequisitionItems
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RequisitionItem"
    post:
      operationId: addRequisitionItem
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisitionItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequisitionItem"
  /requisitions/{id}/items/{itemId}:
    patch:
      operationId: updateRequisitionItem
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisitionItemInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequisitionItem"
    delete:
      operationId: deleteRequisitionItem
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /requisitions/{id}/queries:
    get:
      operationId: listQueries
      tags: [queries]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QueryWithReplies"
    post:
      operationId: raiseQuery
      tags: [queries]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryWithReplies"
  /queries/{queryId}/reply:
    post:
      operationId: replyToQuery
      tags: [queries]
      parameters:
        - name: queryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryReplyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryReplyObject"
  /queries/{queryId}/resolve:
    post:
      operationId: resolveQuery
      tags: [queries]
      parameters:
        - name: queryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveQueryInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryWithReplies"

  # ── Status Updates ───────────────────────────────────────────────────────────
...
  /holidays:
    get:
      operationId: listHolidays
      tags: [holidays]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Holiday"
    post:
      operationId: createHoliday
      tags: [holidays]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HolidayInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holiday"
  /holidays/{id}:
    delete:
      operationId: deleteHoliday
      tags: [holidays]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      parameters:
        - name: project_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    patch:
      operationId: updateAsset
      tags: [assets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [assets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    SessionUser:
      type: object
      required: [id, name, role]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        site_name:
          type: ["string", "null"]
...
    RequisitionItem:
      type: object
      required: [id, requisition_id, item_name, quantity]
      properties:
        id:
          type: integer
        requisition_id:
          type: integer
        item_name:
          type: string
        quantity:
          type: number
        unit:
          type: ["string", "null"]
        reference_no:
          type: ["string", "null"]
        expected_cost:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        remark:
          type: ["string", "null"]
        asset_id:
          type: ["integer", "null"]
        asset_name:
          type: ["string", "null"]
        sort_order:
          type: integer
    RequisitionItemInput:
      type: object
      required: [item_name, quantity]
      properties:
        item_name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        reference_no:
          type: string
        expected_cost:
          type: ["number", "null"]
        description:
          type: string
        remark:
          type: string
        asset_id:
          type: ["integer", "null"]
        asset_name:
          type: string
        sort_order:
          type: integer
    Attachment:
      type: object
      required: [id, requisition_id, filename, original_name]
      properties:
        id:
          type: integer
        requisition_id:
          type: integer
        context:
          type: string
        context_id:
          type: ["integer", "null"]
        filename:
          type: string
        original_name:
          type: string
        mime_type:
          type: ["string", "null"]
        size_bytes:
          type: ["integer", "null"]
        uploaded_by_name:
          type: ["string", "null"]
        uploaded_at:
          type: string
        url:
          type: string
...
    Asset:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: ["string", "null"]
          description: A free-form label like "vessel", "equipment", or "location".
        project_id:
          type: ["integer", "null"]
          description: Null means available on every project.
    AssetInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        type:
          type: string
        project_id:
          type: ["integer", "null"]
...
    QueryWithReplies:
      type: object
      required: [id, requisition_id, raised_by_name, message, is_resolved, created_at]
      properties:
        id:
          type: integer
        requisition_id:
          type: integer
        raised_by_name:
          type: string
        raised_by_role:
          type: ["string", "null"]
        message:
          type: string
        is_resolved:
          type: boolean
        resolved_by_name:
          type: ["string", "null"]
        resolved_at:
          type: ["string", "null"]
        created_at:
          type: string
        replies:
          type: array
          items:
            $ref: "#/components/schemas/QueryReplyObject"
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
    QueryReplyObject:
      type: object
      required: [id, query_id, replied_by_name, message, created_at]
      properties:
        id:
          type: integer
        query_id:
          type: integer
        replied_by_name:
          type: string
        replied_by_role:
          type: ["string", "null"]
        message:
          type: string
        created_at:
          type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
    QueryInput:
      type: object
      required: [raised_by_name, message]
      properties:
        raised_by_name:
          type: string
        raised_by_role:
          type: string
        message:
          type: string
    QueryReplyInput:
      type: object
      required: [replied_by_name, message]
      properties:
        replied_by_name:
          type: string
        replied_by_role:
          type: string
        message:
          type: string
    ResolveQueryInput:
      type: object
      required: [resolved_by_name]
      properties:
        resolved_by_name:
          type: string
...
    Holiday:
      type: object
      required: [id, date, name]
      properties:
        id:
          type: integer
        date:
          type: string
        name:
          type: string
    HolidayInput:
      type: object
      required: [date, name]
      properties:
        date:
          type: string
        name:
          type: string
...